feat(front): create a hook useFileManager to avoid duplicate code in orgaization files and dashboard
